+++ client.h
@@ -12,30 +12,30 @@ class Client
 public:
     Client(const QString& fio,
            const std::function<void(QString)> addNewCard,
-           const std::function<void(QString, unsigned int)> removeCard);
+           const std::function<void(QString, size_t)> removeCard);
     Client(const Client& X);
     Client& operator= (const Client& X);
 
     void AddNewCard();
     void AddNewCard(Card *card);
-    void DeleteCard(const unsigned int& id);
+    void DeleteCard(const size_t& id);
 
-    QList<unsigned int> GetCardIds();
-    double GetCardBalance(const unsigned int& id);
-    void UseCard(const unsigned int& id, const double& value);
+    QList<size_t> GetCardIds();
+    double GetCardBalance(const size_t& id);
+    void UseCard(const size_t& id, const double& value);
 
-    bool IsCardBlocked(const unsigned int& id);
-    void BlockCard(const unsigned int& id);
-    void UnblockCard(const unsigned int& id);
+    bool IsCardBlocked(const size_t& id);
+    void BlockCard(const size_t& id);
+    void UnblockCard(const size_t& id);
 
-    QList<unsigned int> GetCardsInfo();
+    QList<size_t> GetCardsInfo();
 
 private:
     QString m_fio;
-    QMap<unsigned int, Card*> m_cards;
+    QMap<size_t, Card*> m_cards;
 
     std::function<void(QString)> m_addNewCard;
-    std::function<void(QString, unsigned int)> m_removeCard;
+    std::function<void(QString, size_t)> m_removeCard;
 
     // Не доступен
     Client();
